Split PriceCurveSanitizer into two CurveHandlers
